revert part of rev 502. changing total_size from int back to double (int broke large downloads, ie my 9G download was showing up as total size 890MB or so).

+++ src/deluge_core.cpp
@@ -640,7 +640,7 @@ static PyObject *torrent_get_torrent_state(PyObject *self, PyObject *args)
         "pieces",             long(s.pieces),
         "pieces_done",        long(s.num_pieces),
         "block_size",         long(s.block_size),
-        "total_size",         int(i.total_size()),
+        "total_size",         double(i.total_size()),
         "piece_length",       long(i.piece_length()),
         "num_pieces",         long(i.num_pieces()),
         "total_peers",        long(s.num_incomplete != -1? s.num_incomplete : connected_peers),
